IN-111 feat: add Settings page and tabs for Settings Page
