+++ src/common/HOC/withProtection.tsx
@@ -2,41 +2,10 @@
 
 import React, { ReactElement, ReactNode } from 'react'
 
-import {
-  Layout,
-  LayoutWithSidebar,
-  Paths,
-  Tabs,
-  TabsType,
-  Typography,
-} from '@/common'
+import { Layout, LayoutWithSidebar, Paths } from '@/common'
 import authStore from '@/features/auth/model/authStore'
 import { NextPage } from 'next'
 import { useRouter } from 'next/router'
-import GeneralInfo from 'src/pages/general-info'
-
-const tabsData: TabsType[] = [
-  {
-    content: <></> || <GeneralInfo />,
-    id: 'tab1',
-    title: 'General Information',
-  },
-  {
-    content: <Typography variant={'reg16'}>Devices</Typography>,
-    id: 'tab2',
-    title: 'Devices',
-  },
-  {
-    content: <Typography variant={'reg16'}>Account Management</Typography>,
-    id: 'tab3',
-    title: 'Account Management',
-  },
-  {
-    content: <Typography variant={'reg16'}>My payments</Typography>,
-    id: 'tab4',
-    title: 'My payments',
-  },
-]
 
 type NextPageWithLayout<P = {}, IP = P> = {
   getLayout?: (page: ReactElement) => ReactNode
@@ -45,8 +14,7 @@ type NextPageWithLayout<P = {}, IP = P> = {
 export const withProtection =
   <P extends object>(
     PageComponent: NextPageWithLayout<P>,
-    isPublic: boolean = false,
-    isWithTabs?: boolean
+    isPublic: boolean = false
   ): NextPageWithLayout<P> =>
   (props) => {
     const router = useRouter()
@@ -58,12 +26,6 @@ export const withProtection =
     if (authStore.profile) {
       return (
         <LayoutWithSidebar>
-          {isWithTabs && (
-            <Tabs
-              className={'mt-9'}
-              tabsData={tabsData}
-            />
-          )}
           <PageComponent {...props} />
         </LayoutWithSidebar>
       )
